Move genotype_as_bytes to display.py
