+++ sgkit/display.py
@@ -1,7 +1,11 @@
 from typing import Any, Hashable, Mapping, Tuple
 
+import numpy as np
 import pandas as pd
 import xarray as xr
+
+from sgkit.accelerate import numba_guvectorize
+from sgkit.typing import ArrayLike
 
 
 class GenotypeDisplay:
@@ -54,6 +58,88 @@ class GenotypeDisplay:
                     f"<p>{self.shape[0]} rows x {self.shape[1]} columns</p></div>",
                 )
             return self.df._repr_html_()
+
+
+@numba_guvectorize(  # type: ignore
+    [
+        "void(uint8[:], uint8[:], boolean[:], uint8[:], uint8[:])",
+    ],
+    "(b),(),(),(c)->(c)",
+)
+def _format_genotype_bytes(
+    chars: ArrayLike, ploidy: int, phased: bool, _: ArrayLike, out: ArrayLike
+) -> None:  # pragma: no cover
+    ploidy = ploidy[0]
+    sep = 124 if phased[0] else 47  # "|" or "/"
+    chars_per_allele = len(chars) // ploidy
+    slot = 0
+    for slot in range(ploidy):
+        offset_inp = slot * chars_per_allele
+        offset_out = slot * (chars_per_allele + 1)
+        if slot > 0:
+            out[offset_out - 1] = sep
+        for char in range(chars_per_allele):
+            i = offset_inp + char
+            j = offset_out + char
+            val = chars[i]
+            if val == 45:  # "-"
+                if chars[i + 1] == 49:  # "1"
+                    # this is an unknown allele
+                    out[j] = 46  # "."
+                    out[j + 1 : j + chars_per_allele] = 0
+                    break
+                else:
+                    # < -1 indicates a gap
+                    out[j : j + chars_per_allele] = 0
+                    if slot > 0:
+                        # remove separator
+                        out[offset_out - 1] = 0
+                    break
+            else:
+                out[j] = val
+    # shuffle zeros to end
+    c = len(out)
+    for i in range(c):
+        if out[i] == 0:
+            for j in range(i + 1, c):
+                if out[j] != 0:
+                    out[i] = out[j]
+                    out[j] = 0
+                    break
+
+
+def genotype_as_bytes(
+    genotype: ArrayLike,
+    phased: ArrayLike,
+    max_allele_chars: int = 2,
+) -> ArrayLike:
+    """Convert integer encoded genotype calls to (unphased)
+    VCF style byte strings.
+
+    Parameters
+    ----------
+    genotype
+        Genotype call.
+    phased
+        Boolean indicating if genotype is phased.
+    max_allele_chars
+        Maximum number of chars required for any allele.
+        This should include signed sentinel values.
+
+    Returns
+    -------
+    genotype_string
+        Genotype encoded as byte string.
+    """
+    ploidy = genotype.shape[-1]
+    b = genotype.astype("|S{}".format(max_allele_chars))
+    b.dtype = np.uint8
+    n_num = b.shape[-1]
+    n_char = n_num + ploidy - 1
+    dummy = np.empty(n_char, np.uint8)
+    c = _format_genotype_bytes(b, ploidy, phased, dummy)
+    c.dtype = "|S{}".format(n_char)
+    return np.squeeze(c)
 
 
 def truncate(ds: xr.Dataset, max_sizes: Mapping[Hashable, int]) -> xr.Dataset:
